r20503: And one more 64-bit warning

+++ source/tdb/common/io.c
@@ -126,7 +126,7 @@ static int tdb_read(struct tdb_context *tdb, tdb_off_t off, void *buf,
 			tdb->ecode = TDB_ERR_IO;
 			TDB_LOG((tdb, TDB_DEBUG_FATAL,"tdb_read failed at %d "
 				 "len=%d ret=%d (%s) map_size=%d\n",
-				 (int)off, (int)len, ret, strerror(errno),
+				 (int)off, (int)len, (int)ret, strerror(errno),
 				 (int)tdb->map_size));
 			return TDB_ERRCODE(TDB_ERR_IO, -1);
 		}
